fixed cleanup of temp files in test and splitted retrain and test api
